Never specify fallback encoding in output
Some pages don't specify their encoding and we shouldn't do it for them. JSDOM is strict to the standard sniffing algorithm, and so assumes windows-1252. Browsers usually detect correctly that the page meant UTF-8. If the page was really in windows-1252 we never need to specify it anyways, since it is, too, the fallback in the sniffing algorithm of the browser.

The offending page: https://nitter.snopyta.org/the_stevem/status/1599715631472910336 . It's one of those that only uses response headers to set encoding: "content-type: text/html;charset=utf-8"